+++ geopandas/tests/test_geom_methods.py
@@ -900,6 +900,29 @@ class TestGeomMethods:
         assert result.is_valid.all()
 
     @pytest.mark.skipif(
+        not (compat.USE_PYGEOS or compat.USE_SHAPELY_20),
+        reason=("reverse is only implemented for pygeos and shapely >= 2.0"),
+    )
+    def test_reverse(self):
+        expected = GeoSeries(
+            [
+                LineString([(0, 0), (0, 1), (1, 1)]),
+                LineString([(0, 0), (1, 0), (1, 1), (0, 1)]),
+            ]
+        )
+        assert_geoseries_equal(expected, self.g5.reverse())
+
+    @pytest.mark.skipif(
+        (compat.USE_PYGEOS or compat.USE_SHAPELY_20),
+        reason="reverse not implemented for shapely<2",
+    )
+    def test_reverse_not_implemented(self):
+        with pytest.raises(
+            NotImplementedError, match="shapely >= 2.0 or PyGEOS is required"
+        ):
+            self.g5.reverse()
+
+    @pytest.mark.skipif(
         (compat.USE_PYGEOS or compat.USE_SHAPELY_20),
         reason="segmentize keyword introduced in shapely 2.0",
     )
@@ -1831,6 +1854,8 @@ class TestGeomMethods:
                 len(output.explode(ignore_index=True))
                 == len(gs[~(gs.is_empty | gs.isna())]) * size
             )
+        with pytest.warns(FutureWarning, match="The 'seed' keyword is deprecated"):
+            _ = gs.sample_points(size, seed=1)
 
     @pytest.mark.skipif(
         not (compat.USE_PYGEOS or compat.USE_SHAPELY_20),
@@ -1873,3 +1898,49 @@ class TestGeomMethods:
         expected = GeoSeries([LineString([[-1, 0], [-1, 2], [1, 2]])])
         assert_geoseries_equal(expected, oc)
         assert isinstance(oc, GeoSeries)
+
+    @pytest.mark.skipif(
+        compat.SHAPELY_GE_20,
+        reason="remove_repeated_points is implemented for shapely >= 2.0",
+    )
+    def test_remove_repeated_points_not_implemented_shapely_pre2(self):
+        with pytest.raises(
+            NotImplementedError,
+            match=f"shapely >= 2.0 is required, "
+            f"version {shapely.__version__} is installed",
+        ):
+            self.squares.remove_repeated_points()
+
+    @pytest.mark.skipif(
+        not (compat.USE_PYGEOS and compat.SHAPELY_GE_20),
+        reason="remove_repeated_points is only implemented for shapely >= 2.0",
+    )
+    def test_remove_repeated_points_pygeos_set_shapely_installed(self):
+        with pytest.warns(
+            UserWarning,
+            match=(
+                "PyGEOS does not support remove_repeated_points, "
+                "and Shapely >= 2 is installed"
+            ),
+        ):
+            self.g1.remove_repeated_points()
+
+    @pytest.mark.skipif(
+        not compat.USE_SHAPELY_20,
+        reason="remove_repeated_points is only implemented for shapely >= 2.0",
+    )
+    @pytest.mark.parametrize(
+        "geom,expected",
+        [
+            (
+                GeoSeries(LinearRing([(0, 0), (1, 2), (1, 2), (1, 3), (0, 0)])),
+                GeoSeries(LinearRing([(0, 0), (1, 2), (1, 3), (0, 0)])),
+            ),
+            (
+                GeoSeries(Polygon([(0, 0), (0, 0), (1, 0), (1, 1), (1, 0), (0, 0)])),
+                GeoSeries(Polygon([(0, 0), (1, 0), (1, 1), (1, 0), (0, 0)])),
+            ),
+        ],
+    )
+    def test_remove_repeated_points(self, geom, expected):
+        assert_geoseries_equal(expected, geom.remove_repeated_points(tolerance=0.0))