+++ CHANGELOG.md
@@ -7,8 +7,13 @@ PyGEOS are installed.  PyGEOS will continue to be used by default when PyGEOS is
 installed alongside Shapely < 2.0.  Support for PyGEOS and Shapely < 2.0 will be removed
 in GeoPandas 1.0.
 
+API changes:
+
+- ``seed`` keyword in ``sample_points`` is deprecated. Use ``rng`` instead. (#2913).
+
 New methods:
 
+- Added ``remove_repeated_points`` method from shapely to GeoSeries/GeoDataframe (#2940).
 - Added ``frechet_distance()`` method from shapely to GeoSeries/GeoDataframe (#2929).
 - Added ``segmentize`` method from shapely to GeoSeries/GeoDataFrame (#2910).
 - Added ``extract_unique_points`` method from shapely to GeoSeries/GeoDataframe (#2915).
@@ -18,11 +23,12 @@ New methods:
 - Added ``offset_curve`` method from shapely to GeoSeries/GeoDataframe (#2902).
 - Added ``shortest_line`` method from shapely to GeoSeries/GeoDataframe (#2960).
 - Added ``minimum_rotated_rectangle`` method from shapely to GeoSeries/GeoDataframe (#2541).
+- Added ``reverse`` method from shapely to GeoSeries/GeoDataframe (#2988).
 
 New features and improvements:
 
 - Added ``exclusive`` parameter to ``sjoin_nearest`` method for Shapely >= 2.0 (#2877)
-- The ``to_file()`` method will now automatically detect the FlatGeoBuf driver 
+- The ``to_file()`` method will now automatically detect the FlatGeoBuf driver
   for files with the `.fgb` extension (#2958)
 
 Bug fixes: